Preliminary refactoring
