+++ src/TestEngine/testcentric.engine/Services/TestAgency.cs
@@ -5,15 +5,13 @@
 
 #if !NETSTANDARD2_0
 using System;
-using System.IO;
 using System.Threading;
 using System.Diagnostics;
-using System.Text;
 using TestCentric.Common;
 using TestCentric.Engine.Agents;
 using TestCentric.Engine.Internal;
-using TestCentric.Engine.Helpers;
 using NUnit.Engine;
+using TestCentric.Engine.Transports.Remoting;
 
 namespace TestCentric.Engine.Services
 {
